@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -30,9 +30,6 @@ dependencies:
   #A flutter plugin for adapting screen and font size.Let your UI display a reasonable layout on different screen sizes!
   flutter_screenutil: ^5.0.0+2
 
-  #A highly configurable navigation bar with emphasis for the selected item.
-  ff_navigation_bar: ^0.1.4
-
   #A simple image slideshow widget. Mainly intended for image widget, but other widgets can also be used.
   flutter_image_slideshow: ^0.1.0
 
@@ -40,18 +37,42 @@ dependencies:
   extended_sliver: ^2.0.0
 
   #Flutter widget that automatically resizes text to fit perfectly within its bounds.
-  auto_size_text: ^2.1.0
+  auto_size_text: ^3.0.0-nullsafety.0
 
   #A Flutter plugin that provides a Google Maps widget.
-  google_maps_flutter: ^2.0.4
+  google_maps_flutter: ^2.0.6
   geocoding: ^2.0.0
-  geolocator: ^7.0.3
+  geolocator: ^7.1.0
 
   search_page: ^2.0.0+1
 
-# The following adds the Cupertino Icons font to your application.
+  rflutter_alert: ^2.0.2
+
+  flutter_rating_bar: ^4.0.0
+
+  image_picker: ^0.8.0+3
+
+  firebase_auth: ^1.4.1
+
+  firebase_core: ^1.3.0
+
+  google_sign_in: ^5.0.4
+
+  flutter_facebook_auth: ^3.4.1
+
+  firebase_analytics: ^8.1.2
+
+  http: ^0.13.3
+
+  flutter_easyloading: ^3.0.0
+
+  shared_preferences: ^2.0.6
+
+#  cloud_firestore: ^2.2.2
+
+  # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.1
+  cupertino_icons: ^1.0.3
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -89,11 +110,11 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-#  fonts:
-#     - family: Roboto
-#       fonts:
-#         - asset: assets/fonts/Roboto/
-#           style: italic
+  #  fonts:
+  #     - family: Roboto
+  #       fonts:
+  #         - asset: assets/fonts/Roboto/
+  #           style: italic
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
