@@ -1,9 +1,9 @@
 class Address {
   final id;
-  final String addressDetail;
-  final String address;
-  final double long;
-  final double lat;
+  final String? addressDetail;
+  final String? address;
+  final double? long;
+  final double? lat;
 
   Address({this.id, this.addressDetail, this.address, this.long, this.lat});
 }
