@@ -1,6 +1,6 @@
 class Category {
-  String name;
-  String url;
+  String? name;
+  String? url;
 
   Category({this.name, this.url});
 }
