@@ -2,13 +2,13 @@ import 'package:fooddelivery/model/sizes.dart';
 import 'package:fooddelivery/model/toppings.dart';
 
 class Food {
-  final int id;
-  final String name;
-  final String image;
-  final double price;
-  final int restaurantId;
-  final List<Topping> listTopping;
-  final Size size;
+  final int? id;
+  final String? name;
+  final String? image;
+  final double? price;
+  final int? restaurantId;
+  final List<Topping>? listTopping;
+  final Size? size;
 
   Food(
       {this.id,
