@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class FunctionProfile {
   final String name;
-  final Widget widget;
+  final Widget? widget;
 
-  FunctionProfile({@required this.name, @required this.widget});
+  FunctionProfile({required this.name, required this.widget});
 }
