@@ -1,5 +1,5 @@
 class Images {
-  final int id;
+  final int? id;
   final url;
 
   Images({this.id, this.url});
