@@ -1,6 +1,6 @@
 class Notify{
-  final String title;
-  final String content;
+  final String? title;
+  final String? content;
 
   Notify({this.title, this.content});
 }
