@@ -3,14 +3,14 @@ import 'package:fooddelivery/model/restaurants.dart';
 import 'package:fooddelivery/model/users.dart';
 
 class Order {
-  final int id;
-  final Users user;
-  final Restaurants restaurant;
-  final String category;
-  final Map<int,Food> listFood;
-  final double price;
-  final String method;
-  final bool status;
+  final int? id;
+  final Users? user;
+  final Restaurants? restaurant;
+  final String? category;
+  final Map<int,Food>? listFood;
+  final double? price;
+  final String? method;
+  final bool? status;
 
   Order(
       {this.id,
