@@ -1,8 +1,8 @@
 
 class Size {
-  final int id;
-  final String name;
-  final double price;
+  final int? id;
+  final String? name;
+  final double? price;
 
   Size({this.id, this.name, this.price,});
 }
