@@ -1,4 +1,3 @@
-
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:fooddelivery/model/food.dart';
@@ -6,8 +5,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:fooddelivery/model/toppings.dart';
 
 class FoodItem extends StatelessWidget {
-  final Map<int, Food> map;
-  static int i;
+  final Map<int, Food>? map;
+  static int? i;
 
   FoodItem({this.map});
 
@@ -17,7 +16,7 @@ class FoodItem extends StatelessWidget {
     // print('$i dsad');
     return Column(
       children: [
-        for (i = 0; i < map.length; i++)
+        for (i = 0; i! < map!.length; i != i! + 1)
           Container(
             padding:
                 EdgeInsets.only(left: 8.w, right: 8.w, top: 8.h, bottom: 8),
@@ -28,7 +27,7 @@ class FoodItem extends StatelessWidget {
                 Container(
                   width: 50.w,
                   child: Image.network(
-                    map.values.elementAt(i).image,
+                    map!.values.elementAt(i!).image!,
                     width: 50,
                     height: 50,
                     fit: BoxFit.cover,
@@ -36,7 +35,7 @@ class FoodItem extends StatelessWidget {
                 ),
                 Container(
                   width: 24.w,
-                  child: Text('x ' + map.keys.elementAt(i).toString()),
+                  child: Text('x ' + map!.keys.elementAt(i!).toString()),
                 ),
                 Container(
                   width: 230.w,
@@ -44,26 +43,26 @@ class FoodItem extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       AutoSizeText(
-                        map.values.elementAt(i).name,
+                        map!.values.elementAt(i!).name!,
                         style: TextStyle(fontSize: 18.sp),
                       ),
-                      map.values.elementAt(i).size == null
+                      map!.values.elementAt(i!).size == null
                           ? SizedBox()
                           : Text(
-                              'Size ' + map.values.elementAt(i).size.name,
+                              'Size ' + map!.values.elementAt(i!).size!.name!,
                               style: TextStyle(
                                 fontSize: 16.sp,
                               ),
                             ),
-                      map.values.elementAt(i).listTopping == null
+                      map!.values.elementAt(i!).listTopping == null
                           ? Text('')
                           : Column(
                               children: [
                                 for (Topping topping
-                                    in map.values.elementAt(i).listTopping)
+                                    in map!.values.elementAt(i!).listTopping!)
                                   topping.name == null
                                       ? Text('')
-                                      : Text(topping.name),
+                                      : Text(topping.name!),
                               ],
                             )
                     ],
@@ -72,7 +71,7 @@ class FoodItem extends StatelessWidget {
                 Container(
                   width: 70.w,
                   child: Text(
-                    sumPrice(map.values.elementAt(i), map.keys.elementAt(i))
+                    sumPrice(map!.values.elementAt(i!), map!.keys.elementAt(i!))
                             .toStringAsFixed(0) +
                         'đ',
                     textAlign: TextAlign.right,
@@ -85,10 +84,10 @@ class FoodItem extends StatelessWidget {
     );
   }
 
-  double sumPriceTopping(List<Topping> list) {
+  double sumPriceTopping(List<Topping>? list) {
     double sum = 0;
     if (list == null) return sum;
-    for (Topping topping in list) sum += topping.price;
+    for (Topping topping in list) sum += topping.price!;
 
     return sum;
   }
@@ -96,8 +95,8 @@ class FoodItem extends StatelessWidget {
   double priceOfItem(Food food) {
     double sum = 0;
     if (food.size == null)
-      return sum += food.price + sumPriceTopping(food.listTopping);
-    sum += food.price + food.size.price + sumPriceTopping(food.listTopping);
+      return sum += food.price! + sumPriceTopping(food.listTopping);
+    sum += food.price! + food.size!.price! + sumPriceTopping(food.listTopping);
 
     return sum;
   }
