@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 class DeliveryModel {
-  final IconData iconData;
-  final String name;
-  final String address;
+  final IconData? iconData;
+  final String? name;
+  final String? address;
 
   DeliveryModel({
     this.iconData, this.name, this.address
